+++ src/pipelineforge/__init__.py
@@ -1,0 +1,6 @@
+"""
+PipelineForge
+Version: 0.2.0
+"""
+
+__version__ = "0.2.0"